requirements.txt
gradio==4.44.0
numpy==1.24.4
tensorflow==2.13.0
opencv-python-headless==4.8.0.76
Pillow==10.0.0
scikit-learn==1.3.0